Modify generator to use asynchronous generation for efficiency
This modifies the createGIFForTimePoints function to use the
generateCGImagesAsynchronouslyForTimes:completionHandler: method to do
the actual image generation, which is far more efficient and speedy than
copyCGImageAtTime:actualTime:error:.

After this change, the unit tests ran in 40% less time based on my
testing. Other GIFs may experience more or less of a performance bonus,
depending on the video encoding and frames extracted.